Rename error file component and update paths
